requirements.txt
twscrape==0.17.0
kafka-python==2.2.7
pyspark==4.0.0
psycopg2-binary==2.9.10
SQLAlchemy==1.4.54
pandas==2.1.4
numpy==1.26.4
transformers==4.52.4
torch==2.2.2
apache-airflow==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
Sastrawi==1.0.1
pyarrow==20.0.0